refactor(webapp): pay down boy-scout debt in serve-command.ts
Remove shell-layer back-edges from serve-command by relocating preview-minter
hooks into shell/, routing mime detection through shared.detectMimeType, and
importing BrowserAPI via kernel/browser-api instead of cdp/ directly.
